# docs/T033_openapi.yaml
openapi: "3.0.3"
info:
  title: GarudaX Platform API
  description: |
    REST API for the AI Powered Commodity Exchange of Mongolia.
    This gateway translates REST/JSON requests to gRPC calls on backend exchange services.
  version: "1.0.0"
  contact:
    name: GarudaX Platform Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.garudax.mn/api/v1
    description: Production
  - url: https://staging-api.garudax.mn/api/v1
    description: Staging
...
security:
  - BearerAuth: []
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Submit a new order
      operationId: submitOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitOrderRequest"
      responses:
        "201":
          description: Order accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      tags: [Orders]
      summary: List open orders
      operationId: getOpenOrders
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: instrument_id
          in: query
          schema:
            type: string
            format: uuid
        - name: side
          in: query
          schema:
            $ref: "#/components/schemas/Side"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Open orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenOrdersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Orders]
      summary: Cancel all orders for an account
      operationId: cancelAllOrders
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: instrument_id
          in: query
          schema:
            type: string
            format: uuid
        - name: side
          in: query
          schema:
            $ref: "#/components/schemas/Side"
      responses:
        "200":
          description: Cancellation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelAllOrdersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get order details
      operationId: getOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Orders]
      summary: Cancel an order
      operationId: cancelOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Cancellation execution report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionReport"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Orders]
      summary: Modify an order (cancel-replace)
      operationId: modifyOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyOrderRequest"
      responses:
        "200":
          description: Modification execution report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionReport"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Market Data ──────────────────────────────────────────────────────
  /instruments/{instrument_id}/book:
    get:
      tags: [Market Data]
      summary: Get L2 order book snapshot
      operationId: getOrderBook
      security: []
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: depth
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Order book snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBookSnapshot"
  /instruments/{instrument_id}/book/l3:
    get:
      tags: [Market Data]
      summary: Get L3 order book (individual orders)
      operationId: getOrderBookL3
      security: []
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: L3 order book snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBookSnapshotL3"
  /instruments/{instrument_id}/trades/latest:
    get:
      tags: [Market Data]
      summary: Get the most recent trade
      operationId: getLastTrade
      security: []
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Most recent trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"

  # ─── Admin ────────────────────────────────────────────────────────────
  /admin/instruments/{instrument_id}/halt:
    post:
      tags: [Admin]
      summary: Halt trading for an instrument
      operationId: haltInstrument
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Halt confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HaltInstrumentResponse"
  /admin/instruments/{instrument_id}/resume:
    post:
      tags: [Admin]
      summary: Resume trading after halt
      operationId: resumeInstrument
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                trigger_auction:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Resume confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeInstrumentResponse"
  /admin/trades/{trade_id}/bust:
    post:
      tags: [Admin]
      summary: Bust an erroneous trade
      operationId: bustTrade
      parameters:
        - name: trade_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Bust trade result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /admin/instruments/{instrument_id}/circuit-breaker:
    put:
      tags: [Admin]
      summary: Update circuit breaker configuration
      operationId: setCircuitBreaker
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircuitBreakerConfig"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CircuitBreakerConfig"
  /admin/participants/{participant_id}/disable:
    post:
      tags: [Admin]
      summary: Disable a participant from trading
      operationId: disableParticipant
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Disable result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisableParticipantResponse"
  /admin/mass-cancel:
    post:
      tags: [Admin]
      summary: Mass cancel orders
      operationId: massCancel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                instrument_id:
                  type: string
                  format: uuid
                participant_id:
                  type: string
                  format: uuid
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Mass cancel result
          content:
            application/json:
              schema:
                type: object
                properties:
                  cancelled_count:
                    type: integer

  # ─── Clearing ─────────────────────────────────────────────────────────
  /clearing/positions:
    get:
      tags: [Clearing]
      summary: List positions for a participant
      operationId: getPositions
      parameters:
        - name: participant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Position list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Position"
  /clearing/positions/{instrument_id}:
    get:
      tags: [Clearing]
      summary: Get position for specific instrument
      operationId: getPosition
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: participant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Position details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /margin:
    get:
      tags: [Margin]
      summary: Get portfolio margin for authenticated participant
      operationId: getPortfolioMargin
      parameters:
        - name: participant_id
          in: query
          schema:
            type: string
            format: uuid
            description: Defaults to authenticated user's participant_id
      responses:
        "200":
          description: Portfolio margin details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioMargin"
        "404":
          $ref: "#/components/responses/NotFound"
  /margin/calculate:
    post:
      tags: [Margin]
      summary: Trigger margin calculation
      operationId: calculateMargin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                participant_id:
                  type: string
                  format: uuid
              required: [participant_id]
      responses:
        "200":
          description: Margin calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioMargin"
...
  /settlement/cycles/{cycle_id}:
    get:
      tags: [Settlement]
      summary: Get settlement cycle details
      operationId: getSettlementCycle
      parameters:
        - name: cycle_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cycle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementCycle"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Auth ─────────────────────────────────────────────────────────────
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an expired JWT
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
...
  /auth/password/change:
    post:
      tags: [Auth]
      summary: Change password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 12
              required: [current_password, new_password]
      responses:
        "204":
          description: Password changed
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Request password reset
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "202":
          description: Reset email sent (if account exists)

  # ─── Participants / KYC ───────────────────────────────────────────────
  /participants:
    post:
      tags: [Participants]
      summary: Submit KYC application
      operationId: submitApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitApplicationRequest"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KYCApplication"
    get:
      tags: [Participants]
      summary: List KYC applications
      operationId: listApplications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [application_submitted, documents_pending, documents_uploaded, verification_in_progress, screening_in_progress, risk_scoring, manual_review, approved, rejected, suspended, expired]
        - name: type
          in: query
          schema:
            type: string
            enum: [individual, corporate, cooperative, broker, institutional]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Application list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/KYCApplication"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /participants/{participant_id}:
    get:
      tags: [Participants]
      summary: Get KYC application details
      operationId: getApplication
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KYCApplication"
        "404":
          $ref: "#/components/responses/NotFound"
  /participants/{participant_id}/documents:
    post:
      tags: [Participants]
      summary: Upload a KYC document
      operationId: uploadDocument
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document_type:
                  type: string
                  enum: [national_id, passport, proof_of_address, company_registration, beneficial_ownership, financial_statements, broker_license, board_resolution, tax_registration, cooperative_membership]
                file:
                  type: string
                  format: binary
              required: [document_type, file]
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    get:
      tags: [Participants]
      summary: List documents for a participant
      operationId: listDocuments
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
  /participants/{participant_id}/approve:
    post:
      tags: [Participants]
      summary: Approve a KYC application
      operationId: approveApplication
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
      responses:
        "200":
          description: Application approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KYCApplication"
  /participants/{participant_id}/reject:
    post:
      tags: [Participants]
      summary: Reject a KYC application
      operationId: rejectApplication
      parameters:
        - name: participant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KYCApplication"

  # ─── Screening ────────────────────────────────────────────────────────
...
  /screening/{screening_id}:
    get:
      tags: [Screening]
      summary: Get screening result
      operationId: getScreeningResult
      parameters:
        - name: screening_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Screening result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningResult"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SubmitOrderRequest:
      type: object
      properties:
        client_order_id:
          type: string
        instrument_id:
          type: string
          format: uuid
        side:
          $ref: "#/components/schemas/Side"
        order_type:
          type: string
          enum: [limit, market, stop_limit, stop_market]
        time_in_force:
          type: string
          enum: [day, gtc, gtd, ioc, fok]
          default: day
        price:
          type: string
          description: Decimal string, required for limit orders
        stop_price:
          type: string
          description: Decimal string, required for stop orders
        quantity:
          type: integer
          minimum: 1
          description: Quantity in lots
        expire_at:
          type: string
          format: date-time
          description: Required for GTD orders
        stp_mode:
          type: string
          enum: [cancel_newest, cancel_oldest, cancel_both]
          default: cancel_newest
      required: [instrument_id, side, order_type, quantity]
...